==> notebooks/02_whisper_tiny_training.ipynb <==
!pip install -q transformers accelerate datasets torchaudio librosa


import os
import pandas as pd
import numpy as np
from datasets import Dataset


metadata_path = "../data/final_metadata.csv"
df = pd.read_csv(metadata_path)

print("Total samples:", len(df))
df.head()


from transformers import WhisperProcessor

processor = WhisperProcessor.from_pretrained(
    "openai/whisper-tiny",
    language=None,
    task="transcribe"
)

print("Whisper processor loaded")


dataset = Dataset.from_pandas(df)

print(dataset)


import librosa
import torch

def prepare_example(example):
    import librosa
    from transformers import WhisperProcessor

    # Load processor INSIDE worker
    processor = WhisperProcessor.from_pretrained(
        "openai/whisper-tiny",
        language=None,
        task="transcribe"
    )

    # 1. Load audio
    audio, sr = librosa.load(example["file_path"], sr=16000)

    # 2. Audio → Whisper features
    input_features = processor.feature_extractor(
        audio,
        sampling_rate=16000
    ).input_features[0]

    # 3. Tokenize transcript
    labels = processor.tokenizer(
        example["transcript"],
        truncation=True
    ).input_ids

    return {
        "input_features": input_features,
        "labels": labels
    }


processed_dataset = dataset.map(
    prepare_example,
    remove_columns=dataset.column_names,
    num_proc=1
)

print(processed_dataset)



processed_dataset.save_to_disk("../data/processed_whisper_dataset")
