==> notebooks/03_whisper_inference.ipynb <==
import torch
import librosa
from transformers import WhisperProcessor, WhisperForConditionalGeneration


processor = WhisperProcessor.from_pretrained(
    "openai/whisper-tiny",
    task="transcribe"
)

model = WhisperForConditionalGeneration.from_pretrained(
    "openai/whisper-tiny"
)

model.eval()

print("Whisper Tiny loaded successfully")


audio_path = "../data/english/processed_clips/afrikaans1.wav"
# ya
# audio_path = "../data/common_voice_hindi/processed_clips/common_voice_hi_26008353.wav"

audio, sr = librosa.load(audio_path, sr=16000)

print("Audio loaded:", audio_path)


inputs = processor(
    audio,
    sampling_rate=16000,
    return_tensors="pt"
)


with torch.no_grad():
    predicted_ids = model.generate(inputs["input_features"])

transcription = processor.batch_decode(
    predicted_ids,
    skip_special_tokens=True
)[0]

print("Transcription:")
print(transcription)


def transcribe_audio(audio_path):
    audio, sr = librosa.load(audio_path, sr=16000)

    inputs = processor(
        audio,
        sampling_rate=16000,
        return_tensors="pt"
    )

    with torch.no_grad():
        predicted_ids = model.generate(inputs["input_features"])

    text = processor.batch_decode(
        predicted_ids,
        skip_special_tokens=True
    )[0]

    return text


result = transcribe_audio(
    "../data/english/processed_clips/afrikaans1.wav"
)
print(result)
